+++ packages/console/src/components/PaymentOverdueModal/index.tsx
@@ -1,14 +1,14 @@
-import { useContext, useMemo, useState } from 'react';
+import { useContext, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import ReactModal from 'react-modal';
 
 import { contactEmailLink } from '@/consts';
-import { isCloud } from '@/consts/env';
 import { TenantsContext } from '@/contexts/TenantsProvider';
 import Button from '@/ds-components/Button';
 import FormField from '@/ds-components/FormField';
 import InlineNotification from '@/ds-components/InlineNotification';
 import ModalLayout from '@/ds-components/ModalLayout';
+import useOverdueInvoices from '@/hooks/use-overdue-invoices';
 import useSubscribe from '@/hooks/use-subscribe';
 import modalStyles from '@/scss/modal.module.scss';
 
@@ -23,45 +23,13 @@ function PaymentOverdueModal() {
 
   const { visitManagePaymentPage } = useSubscribe();
   const [isActionLoading, setIsActionLoading] = useState(false);
-
   const [hasClosed, setHasClosed] = useState(false);
 
-  const isEnterprise = currentTenant?.subscription.isEnterprisePlan;
+  const { hasOverdueInvoices } = useOverdueInvoices();
 
   const handleCloseModal = () => {
     setHasClosed(true);
   };
-
-  const hasOverdueInvoices = useMemo(() => {
-    if (!isCloud || openInvoices.length === 0) {
-      return false;
-    }
-
-    // Enterprise tenants' invoices are manually paid and has a due date in the future
-    // Should show the modal only if the invoices are overdue.
-    // Keep this for now as some legacy invoices data might not have the `collectionMethod` and `dueDate` fields
-    // TODO: this is a temporary fix to hide the modal for enterprise tenants. Remove this after all legacy invoices are handled.
-    if (isEnterprise) {
-      return false;
-    }
-
-    return openInvoices.some(({ dueDate, collectionMethod }) => {
-      switch (collectionMethod) {
-        // For automatic collection method, consider open invoices always overdue
-        case 'charge_automatically': {
-          return true;
-        }
-        // For manual collection method, consider invoices overdue if past due date
-        case 'send_invoice': {
-          return dueDate && dueDate < new Date();
-        }
-        default: {
-          // For legacy invoices without collection method, consider them overdue
-          return true;
-        }
-      }
-    });
-  }, [isEnterprise, openInvoices]);
 
   if (!hasOverdueInvoices || hasClosed) {
     return null;
